Fix bug in Line::addHead
