+++ HyperMetro/task/src/metro/command/Remove.java
@@ -4,8 +4,8 @@ import metro.entity.Line;
 import metro.entity.Metro;
 import metro.ui.UserInterface;
 
-public class Append extends ManageStation {
-    public Append(final Metro metro, final UserInterface ui) {
+public class Remove extends ManageStation {
+    public Remove(final Metro metro, final UserInterface ui) {
         super(metro, ui, Line::append);
     }
