+++ HyperMetro/task/src/metro/services/Request.java
@@ -15,6 +15,6 @@ public class Request implements Runnable {
 
     @Override
     public void run() {
-        command.execute(params);
+        command.accept(params);
     }
 }
